Migrado do Flow para o Typescript
