+++ package.json
@@ -16,6 +16,10 @@
     "@react-native-community/masked-view": "^0.1.6",
     "@react-navigation/native": "^5.0.8",
     "@react-navigation/stack": "^5.1.0",
+    "@types/jest": "^25.1.4",
+    "@types/react": "^16.9.23",
+    "@types/react-native": "^0.61.20",
+    "@types/react-test-renderer": "^16.9.2",
     "axios": "^0.19.2",
     "base-64": "^0.1.0",
     "firebase": "^7.9.3",
@@ -37,13 +41,16 @@
     "redux": "^4.0.5",
     "redux-persist": "^6.0.0",
     "redux-saga": "^1.1.3",
-    "styled-components": "^5.0.1"
+    "styled-components": "^5.0.1",
+    "typescript": "^3.8.3"
   },
   "devDependencies": {
     "@babel/core": "^7.8.6",
     "@babel/preset-flow": "^7.8.3",
     "@babel/runtime": "^7.8.4",
     "@react-native-community/eslint-config": "^0.0.7",
+    "@types/react-redux": "^7.1.7",
+    "@types/ws": "^7.2.2",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^25.1.0",
     "eslint": "^6.8.0",
