+++ src/model/User/index.ts
@@ -1,0 +1,17 @@
+export default class User  {
+
+  constructor(id: string, name:string, email: string, password: string) {
+    this.id = id;
+    this.name = name;
+    this.email = email;
+    this.password = password;
+  }
+
+  id: string;
+
+  name: string;
+
+  email: string;
+
+  password: string;
+}
