+++ src/services/api.js
@@ -1,4 +1,5 @@
 import * as firebase from 'firebase';
+
 import 'firebase/firestore';
 import config from '../config/firebaseConfig';
 
@@ -6,8 +7,11 @@ if (!firebase.apps.length) {
   firebase.initializeApp(config);
 }
 
-export default firebase;
+// ;export default firebase;
 export const firebaseFirestore = firebase.firestore();
+/*
 export const firebaseAuth = firebase.auth();
 export const firebaseDatabase = firebase.database();
 export const firebaseStorage = firebase.storage();
+
+*/
