--- src/services/authService/index.js
@@ -1,41 +1,0 @@
-/* eslint-disable class-methods-use-this */
-// @flow
-import { sha256 } from 'js-sha256';
-import { useDispatch } from 'react-redux';
-import User from '../../model/User';
-
-import { signOut as signOutAction } from '../../store/modules/auth/actions';
-
-import { firebaseFirestore } from '../api';
-
-export class AuthService {
-  signIn(user: User): Promise<User> {
-    return new Promise((resolve, reject) => {
-      const userRef = firebaseFirestore
-        .collection('usuarios')
-        .where('email', '==', user.email);
-
-      userRef.get().then(snapshot => {
-        let filteredUser: User | null = null;
-        snapshot.forEach(doc => {
-          filteredUser = new User(doc.data().email, doc.data().senha);
-        });
-
-        if (filteredUser) {
-          if (filteredUser.password === sha256(user.password)) {
-            filteredUser.password = sha256(filteredUser.password);
-            resolve(filteredUser);
-          } else {
-            reject(new Error('Password does not match'));
-          }
-        } else {
-          reject(new Error('Email does not exist'));
-        }
-      });
-    });
-  }
-
-  signOut(): void {
-    useDispatch(signOutAction());
-  }
-}
