+++ src/store/modules/auth/sagas.js
@@ -1,3 +1,5 @@
+/* eslint-disable import/no-unresolved */
+/* eslint-disable import/extensions */
 import { Alert } from 'react-native';
 import { takeLatest, call, put, all } from 'redux-saga/effects';
 import User from '../../../model/User';
@@ -8,12 +10,9 @@ import { signInSuccess, signFailure } from './actions';
 export function* signIn({ payload }) {
   try {
     const { email, password } = payload;
-
-    const user = new User(email, password);
+    const user = new User('', '', email, password);
     const auth = new AuthService();
-
     const userSigned = yield call(auth.signIn, user);
-
     yield put(signInSuccess(userSigned));
   } catch (error) {
     Alert.alert(
